Factor out pod identity related values into identity.rs.
