+++ src/scheduler.rs
@@ -20,146 +20,26 @@
 //! that pod id's are "stable" and have a semantic known to the calling operator.
 //!
 //!
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::cell::RefCell;
+use std::collections::hash_map::DefaultHasher;
+use std::collections::{BTreeMap, HashSet};
 use std::fmt::{Debug, Display, Formatter};
+use std::ops::DerefMut;
+
+use kube::api::Resource;
+use serde::de::DeserializeOwned;
+use serde_json::json;
 
 use crate::client::Client;
-use crate::error::OperatorResult;
-use crate::labels;
+use crate::error::{Error, OperatorResult};
+use crate::identity::{NodeIdentity, PodIdentity, PodIdentityFactory, PodToNodeMapping};
 use crate::role_utils::EligibleNodesForRoleAndGroup;
-use k8s_openapi::api::core::v1::{Node, Pod};
-use kube::api::Resource;
-use schemars::JsonSchema;
-use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-use std::cell::RefCell;
-use std::collections::btree_map::Iter;
-use std::collections::hash_map::DefaultHasher;
-use std::convert::TryFrom;
-use std::hash::{Hash, Hasher};
-use std::ops::DerefMut;
-use tracing::{error, warn};
-
-const DEFAULT_NODE_NAME: &str = "<no-nodename-set>";
-const SEMICOLON: &str = ";";
-const REQUIRED_LABELS: [&str; 4] = [
-    labels::APP_NAME_LABEL,
-    labels::APP_INSTANCE_LABEL,
-    labels::APP_COMPONENT_LABEL,
-    labels::APP_ROLE_GROUP_LABEL,
-];
-
-#[derive(Debug, thiserror::Error, PartialEq)]
-pub enum Error {
-    #[error(
-        "Not enough nodes [{number_of_nodes}] available to schedule pods [{number_of_pods}]. Unscheduled pods: {unscheduled_pods:?}."
-    )]
-    NotEnoughNodesAvailable {
-        number_of_nodes: usize,
-        number_of_pods: usize,
-        unscheduled_pods: Vec<PodIdentity>,
-    },
-
-    #[error("PodIdentity could not be parsed: {pod_id:?}. This should not happen. Please open a ticket.")]
-    PodIdentityNotParseable { pod_id: String },
-
-    #[error("Cannot build PodIdentity from Pod without labels. Missing labels: {0:?}")]
-    PodWithoutLabelsNotSupported(Vec<&'static str>),
-
-    #[error("Cannot build NodeIdentity from node without name.")]
-    NodeWithoutNameNotSupported,
-
-    #[error("Cannot construct PodIdentity from empty id field.")]
-    PodIdentityFieldEmpty,
-}
-
-/// Returns a Vec of pod identities according to the replica per role+group pair from `eligible_nodes`.
-///
-/// The `id` field is in the range from one (1) to the number of replicas per role+group. If no replicas
-/// are defined, then the range goes from one (1) to the number of eligible groups.
-///
-/// Given a role with two groups, one two replicas and the other with three replicas, the
-/// generated `id` fields of the pod identities are counted as follows:
-///
-/// ```yaml
-/// role_1:
-///     - group_1:
-///         - id: 1
-///         - id: 2
-///     - group_2:
-///         - id: 1
-///         - id: 2
-///         - id: 3
-/// ```
-///
-/// *NOTE* This function is tightly coupled with [`PodToNodeMapping::try_from_pods`]! If you change it's
-/// implementation you also have to update that one.
-///
-/// # Arguments
-/// * `app_name` - Application name
-/// * `instance` - Service instance
-/// * `eligible_nodes` - Eligible nodes grouped by role and groups.
-pub fn generate_ids(
-    app_name: &str,
-    instance: &str,
-    eligible_nodes: &EligibleNodesForRoleAndGroup,
-) -> Vec<PodIdentity> {
-    let mut generated_ids = vec![];
-    for (role_name, groups) in eligible_nodes {
-        for (group_name, eligible_nodes) in groups {
-            let ids_per_group = eligible_nodes
-                .replicas
-                .map(usize::from)
-                .unwrap_or_else(|| eligible_nodes.nodes.len());
-            for id in 1..ids_per_group + 1 {
-                generated_ids.push(PodIdentity {
-                    app: app_name.to_string(),
-                    instance: instance.to_string(),
-                    role: role_name.clone(),
-                    group: group_name.clone(),
-                    id: id.to_string(),
-                });
-            }
-        }
-    }
-
-    // Sort the result to make testing predictable. Otherwise the for-loop above is not
-    // guaranteed to preserve insertion order so the tests might fail at random.
-    generated_ids.sort_by(|a, b| a.partial_cmp(b).unwrap());
-    generated_ids
-}
-
-#[derive(
-    Clone, Debug, Default, Deserialize, Eq, Hash, JsonSchema, Ord, PartialEq, PartialOrd, Serialize,
-)]
-#[serde(rename_all = "camelCase")]
-#[serde(try_from = "String")]
-#[serde(into = "String")]
-pub struct PodIdentity {
-    app: String,
-    instance: String,
-    role: String,
-    group: String,
-    id: String,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Eq, Hash, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct NodeIdentity {
-    pub name: String,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PodToNodeMapping {
-    mapping: BTreeMap<PodIdentity, NodeIdentity>,
-}
+use k8s_openapi::api::core::v1::Pod;
 
 pub trait PodPlacementHistory {
-    fn find(&self, pod_id: &PodIdentity) -> Option<&NodeIdentity>;
-    fn find_all(&self, pods: &[&PodIdentity]) -> Vec<Option<&NodeIdentity>> {
-        pods.iter().map(|p| self.find(*p)).collect()
+    fn find(&self, pod_id: &PodIdentity) -> Option<NodeIdentity>;
+    fn find_all(&self, pods: &[PodIdentity]) -> Vec<Option<NodeIdentity>> {
+        pods.iter().map(|p| self.find(p)).collect()
     }
 
     fn update(&mut self, pod_id: &PodIdentity, node_id: &NodeIdentity);
@@ -195,14 +75,14 @@ pub trait Scheduler {
     /// Implementations may return an error if not all pods can be mapped to nodes.
     ///
     /// # Arguments:
-    /// * `pods` - The list of desired pods. Should contain both already mapped as well as new pods.
+    /// * `pod_id_factory` - A factory object for all pod identities required by the service.
     /// * `nodes` - Currently available nodes in the system grouped by role and group.
-    /// * `mapped_pods` - Pods that are already mapped to nodes.
+    /// * `pods` - Pods that are already mapped to nodes.
     fn schedule(
         &mut self,
-        pods: &[PodIdentity],
+        pod_id_factory: &dyn PodIdentityFactory,
         nodes: &RoleGroupEligibleNodes,
-        mapped_pods: &PodToNodeMapping,
+        pods: &[Pod],
     ) -> SchedulerResult<SchedulerState>;
 }
 
@@ -218,8 +98,8 @@ pub trait PodPlacementStrategy {
     /// * `preferred_nodes` - Optional nodes to prioritize during placement (if not None)
     fn place(
         &self,
-        pods: &[&PodIdentity],
-        preferred_nodes: &[Option<&NodeIdentity>],
+        pods: &[PodIdentity],
+        preferred_nodes: &[Option<NodeIdentity>],
     ) -> Vec<Option<NodeIdentity>>;
 }
 
@@ -312,190 +192,9 @@ impl SchedulerState {
     }
 }
 
-impl TryFrom<&Pod> for NodeIdentity {
-    type Error = Error;
-    fn try_from(p: &Pod) -> SchedulerResult<Self> {
-        let node_name = p
-            .spec
-            .as_ref()
-            .map(|s| s.node_name.as_ref())
-            .ok_or(Error::NodeWithoutNameNotSupported)?;
-
-        Ok(NodeIdentity {
-            name: node_name.unwrap().clone(),
-        })
-    }
-}
-
-impl TryFrom<String> for PodIdentity {
-    type Error = Error;
-    fn try_from(s: String) -> Result<Self, Error> {
-        let split = s.split(SEMICOLON).collect::<Vec<&str>>();
-        if split.len() != 5 {
-            return Err(Error::PodIdentityNotParseable { pod_id: s });
-        }
-        Ok(PodIdentity::new(
-            split[0], split[1], split[2], split[3], split[4],
-        ))
-    }
-}
-
-impl From<PodIdentity> for String {
-    fn from(pod_id: PodIdentity) -> Self {
-        [
-            pod_id.app,
-            pod_id.instance,
-            pod_id.role,
-            pod_id.group,
-            pod_id.id,
-        ]
-        .join(SEMICOLON)
-    }
-}
-
-impl PodToNodeMapping {
-    /// Returns a `PodToNodeMapping` where pod ids are built from (expected) pod labels and self generated counter.
-    ///
-    /// Returns an `Err` if any of the required labels is not present.
-    ///
-    /// This is intended for pods that do not maintain state on the nodes where they are scheduled,
-    /// such as Spark masters or workers. For ZooKeeper see [`Self::try_from_pods_and_id_label`].
-    ///
-    /// *NOTE* This function assumes that the operator has also generated id labels by using a counter
-    /// starting as one (1) such as [`generate_ids`] in this module. If this was not the case, you
-    /// should not use this function to construct [`PodToNodeMapping`]s from pods! Any change here
-    /// needs to be implemented in [`generate_ids`] too!
-    pub fn try_from_pods(pods: &[Pod]) -> SchedulerResult<PodToNodeMapping> {
-        let mut result = PodToNodeMapping::default();
-        let mut label_count: HashMap<String, usize> = HashMap::new();
-        for p in pods {
-            let index = label_count
-                .entry(PodIdentity::labels(p)?)
-                .and_modify(|e| *e += 1)
-                .or_insert(1);
-            let index_str = (*index).to_string();
-            result.insert(
-                PodIdentity::try_from_pod_and_id(p, &index_str)?,
-                NodeIdentity::try_from(p)?,
-            );
-        }
-        Ok(result)
-    }
-
-    /// Returns a `PodToNodeMapping` where pod ids are built from (expected) pod labels including an id label
-    /// that must be present.
-    ///
-    /// Returns an `Err` if any of the required labels is not present.
-    ///
-    /// This is intended for pods that explicitly maintain state on the nodes where they are scheduled,
-    /// such as ZooKeeper pods. For services where this is not the case, such as Spark masters or workers,
-    /// see [`Self::try_from_pods`].
-    pub fn try_from_pods_and_id_label(
-        pods: &[Pod],
-        id_label_name: &'static str,
-    ) -> SchedulerResult<PodToNodeMapping> {
-        let mut pod_node_mapping = PodToNodeMapping::default();
-
-        for pod in pods {
-            match &pod.metadata.labels {
-                Some(labels) => {
-                    let id = labels
-                        .get(id_label_name)
-                        .ok_or_else(|| Error::PodWithoutLabelsNotSupported(vec![id_label_name]))?;
-                    pod_node_mapping.insert(
-                        PodIdentity::try_from_pod_and_id(pod, id.as_ref())?,
-                        NodeIdentity::try_from(pod)?,
-                    );
-                }
-                None => {
-                    return Err(Error::PodWithoutLabelsNotSupported(
-                        REQUIRED_LABELS.to_vec(),
-                    ))
-                }
-            }
-        }
-        Ok(pod_node_mapping)
-    }
-
-    pub fn iter(&self) -> Iter<'_, PodIdentity, NodeIdentity> {
-        self.mapping.iter()
-    }
-
-    pub fn get_filtered(&self, role: &str, group: &str) -> BTreeMap<PodIdentity, NodeIdentity> {
-        let mut filtered = BTreeMap::new();
-        for (pod_id, node_id) in &self.mapping {
-            if pod_id.role == *role && pod_id.group == *group {
-                filtered.insert(pod_id.clone(), node_id.clone());
-            }
-        }
-        filtered
-    }
-
-    pub fn get(&self, pod_id: &PodIdentity) -> Option<&NodeIdentity> {
-        self.mapping.get(pod_id)
-    }
-
-    pub fn insert(&mut self, pod_id: PodIdentity, node_id: NodeIdentity) -> Option<NodeIdentity> {
-        self.mapping.insert(pod_id, node_id)
-    }
-
-    pub fn filter(&self, id: &PodIdentity) -> Vec<NodeIdentity> {
-        self.mapping
-            .iter()
-            .filter_map(|(pod_id, node_id)| {
-                if pod_id.app == id.app
-                    && pod_id.instance == id.instance
-                    && pod_id.role == id.role
-                    && pod_id.group == id.group
-                {
-                    Some(node_id.clone())
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    pub fn merge(&self, other: &Self) -> Self {
-        let mut temp = self.mapping.clone();
-        temp.extend(other.clone().mapping);
-        PodToNodeMapping { mapping: temp }
-    }
-
-    /// Return true if the `node` is already mapped by a pod from `role` and `group`.
-    pub fn mapped_by(&self, node: &NodeIdentity, role: &str, group: &str) -> bool {
-        for (pod_id, mapped_node) in self.mapping.iter() {
-            if node == mapped_node && pod_id.role == role && pod_id.group == group {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Given `pods` return all that are not mapped.
-    pub fn missing<'a>(&self, pods: &'a [PodIdentity]) -> Vec<&'a PodIdentity> {
-        let mut result = vec![];
-        for p in pods {
-            if !self.mapping.contains_key(p) {
-                result.push(p)
-            }
-        }
-        result
-    }
-
-    #[cfg(test)]
-    pub fn new(map: Vec<(PodIdentity, NodeIdentity)>) -> Self {
-        let mut result = BTreeMap::new();
-        for (p, n) in map {
-            result.insert(p.clone(), n.clone());
-        }
-        PodToNodeMapping { mapping: result }
-    }
-}
-
 impl PodPlacementHistory for K8SUnboundedHistory<'_> {
-    fn find(&self, pod_id: &PodIdentity) -> Option<&NodeIdentity> {
-        self.history.get(pod_id)
+    fn find(&self, pod_id: &PodIdentity) -> Option<NodeIdentity> {
+        self.history.get(pod_id).cloned()
     }
 
     ///
@@ -504,7 +203,7 @@ impl PodPlacementHistory for K8SUnboundedHistory<'_> {
     fn update(&mut self, pod_id: &PodIdentity, node_id: &NodeIdentity) {
         if let Some(history_node_id) = self.find(pod_id) {
             // found but different
-            if history_node_id != node_id {
+            if history_node_id != *node_id {
                 self.history.insert(pod_id.clone(), node_id.clone());
                 self.modified = true;
             }
@@ -519,17 +218,6 @@ impl PodPlacementHistory for K8SUnboundedHistory<'_> {
 impl Display for NodeIdentity {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.name)
-    }
-}
-
-impl From<Node> for NodeIdentity {
-    fn from(node: Node) -> Self {
-        NodeIdentity {
-            name: node
-                .metadata
-                .name
-                .unwrap_or_else(|| DEFAULT_NODE_NAME.to_string()),
-        }
     }
 }
 
@@ -559,27 +247,27 @@ where
     ///
     /// The nodes where unscheduled pods are mapped are selected by the configured strategy.
     /// # Arguments:
-    /// * `pods` - all pod ids required by the service.
+    /// * `pod_id_factory` - a provider for all pod ides required by the system.
     /// * `nodes` - all eligible nodes available in the system
-    /// * `mapped_pods` - existing pod to node mapping
+    /// * `pods` - existing pods that are mapped to nodes.
     fn schedule(
         &mut self,
-        pods: &[PodIdentity],
+        pod_id_factory: &dyn PodIdentityFactory,
         nodes: &RoleGroupEligibleNodes,
-        mapped_pods: &PodToNodeMapping,
+        pods: &[Pod],
     ) -> SchedulerResult<SchedulerState> {
-        let unscheduled_pods = mapped_pods.missing(pods);
+        let unscheduled_pods = pod_id_factory.missing(pods)?;
         let history_nodes = self.history.find_all(unscheduled_pods.as_slice());
-
-        let strategy = self.strategy(nodes, mapped_pods);
+        let mapping = pod_id_factory.try_mapping(pods)?;
+        let strategy = self.strategy(nodes, &mapping);
         let selected_nodes = strategy.place(unscheduled_pods.as_slice(), history_nodes.as_slice());
 
         self.update_history_and_result(
-            unscheduled_pods,
-            selected_nodes.to_vec(),
+            unscheduled_pods.as_slice(),
+            selected_nodes.as_slice(),
             pods.len(),
             nodes.count_unique_node_ids(),
-            mapped_pods,
+            &mapping,
         )
     }
 }
@@ -612,8 +300,8 @@ where
     /// * `current_mapping` - existing pod to node mapping
     fn update_history_and_result(
         &mut self,
-        pods: Vec<&PodIdentity>,
-        nodes: Vec<Option<NodeIdentity>>,
+        pods: &[PodIdentity],
+        nodes: &[Option<NodeIdentity>],
         number_of_pods: usize,
         number_of_nodes: usize,
         current_mapping: &PodToNodeMapping,
@@ -622,7 +310,7 @@ where
         let mut result_err = vec![];
         let mut result_ok = BTreeMap::new();
 
-        for (pod, opt_node) in pods.iter().zip(&nodes) {
+        for (pod, opt_node) in pods.iter().zip(nodes) {
             match opt_node {
                 Some(node) => {
                     // Found a node to schedule on so update the result
@@ -630,7 +318,7 @@ where
                     // and update the history if needed.
                     self.history.update(pod, node);
                 }
-                None => result_err.push((*pod).clone()), // No node available for this pod
+                None => result_err.push(format!("{:?}", (*pod).clone())), // No node available for this pod
             }
         }
 
@@ -681,7 +369,7 @@ impl RoleGroupEligibleNodes {
     fn preferred_node_or<F>(
         &self,
         pod: &PodIdentity,
-        preferred: Option<&NodeIdentity>,
+        preferred: Option<NodeIdentity>,
         default: F,
     ) -> Option<NodeIdentity>
     where
@@ -689,12 +377,12 @@ impl RoleGroupEligibleNodes {
     {
         match self
             .node_set
-            .get(&pod.role)
-            .and_then(|role| role.get(&pod.group))
+            .get(&pod.role().to_string())
+            .and_then(|role| role.get(&pod.group().to_string()))
         {
             Some(nodes) if !nodes.is_empty() => {
                 if let Some(node_id) = preferred {
-                    let tmp = nodes.iter().find(|n| *n == node_id);
+                    let tmp = nodes.iter().find(|n| n == &&node_id);
                     if tmp.is_some() {
                         return tmp.cloned();
                     }
@@ -709,7 +397,7 @@ impl RoleGroupEligibleNodes {
     fn preferred_node_or_last(
         &self,
         pod: &PodIdentity,
-        preferred: Option<&NodeIdentity>,
+        preferred: Option<NodeIdentity>,
     ) -> Option<NodeIdentity> {
         self.preferred_node_or(pod, preferred, |_pod, nodes| nodes.last().cloned())
     }
@@ -751,24 +439,21 @@ impl<'a> GroupAntiAffinityStrategy<'a> {
     pub fn select_node_for_pod(
         &self,
         pod_id: &PodIdentity,
-        preferred_node: Option<&NodeIdentity>,
+        preferred_node: Option<NodeIdentity>,
     ) -> Option<NodeIdentity> {
         let mut borrowed_nodes = self.eligible_nodes.borrow_mut();
 
         // Find a node to schedule on (it might be the node from history)
-        while let Some(next_node) = borrowed_nodes.preferred_node_or_last(pod_id, preferred_node) {
-            borrowed_nodes.remove_eligible_node(
-                &next_node,
-                pod_id.role.as_str(),
-                pod_id.group.as_str(),
-            );
+        while let Some(next_node) =
+            borrowed_nodes.preferred_node_or_last(pod_id, preferred_node.clone())
+        {
+            borrowed_nodes.remove_eligible_node(&next_node, pod_id.role(), pod_id.group());
 
             // check that the node is not already in use *by a pod from the same role+group*
-            if !self.existing_mapping.mapped_by(
-                &next_node,
-                pod_id.role.as_str(),
-                pod_id.group.as_str(),
-            ) {
+            if !self
+                .existing_mapping
+                .mapped_by(&next_node, pod_id.role(), pod_id.group())
+            {
                 return Some(next_node);
             }
         }
@@ -782,158 +467,17 @@ impl PodPlacementStrategy for GroupAntiAffinityStrategy<'_> {
     /// might not reflect the reality between calls.
     fn place(
         &self,
-        pods: &[&PodIdentity],
-        preferred_nodes: &[Option<&NodeIdentity>],
+        pods: &[PodIdentity],
+        preferred_nodes: &[Option<NodeIdentity>],
     ) -> Vec<Option<NodeIdentity>> {
         assert_eq!(pods.len(), preferred_nodes.len());
         pods.iter()
             .zip(preferred_nodes.iter())
-            .map(|(pod, preferred_node)| self.select_node_for_pod(*pod, *preferred_node))
+            .map(|(pod, preferred_node)| self.select_node_for_pod(pod, preferred_node.clone()))
             .collect()
     }
 }
 
-impl PodIdentity {
-    pub fn new(app: &str, instance: &str, role: &str, group: &str, id: &str) -> Self {
-        Self::warn_forbidden_char(app, instance, role, group, id);
-        PodIdentity {
-            app: app.to_string(),
-            instance: instance.to_string(),
-            role: role.to_string(),
-            group: group.to_string(),
-            id: id.to_string(),
-        }
-    }
-
-    /// Returns a string with all pod labels required by the [`PodIdentity`] joined with comma.
-    /// If any required pod labels are missing, returns a [`Error::PodWithoutLabelsNotSupported`].
-    pub fn labels(pod: &Pod) -> SchedulerResult<String> {
-        if pod.metadata.labels.is_none() {
-            return Err(Error::PodWithoutLabelsNotSupported(
-                REQUIRED_LABELS.to_vec(),
-            ));
-        }
-
-        let mut result: Vec<String> = vec![];
-
-        let pod_labels = &pod.metadata.labels.as_ref().unwrap();
-        let mut missing_labels = Vec::with_capacity(REQUIRED_LABELS.len());
-        for label_name in REQUIRED_LABELS {
-            match pod_labels.get(label_name).cloned() {
-                Some(value) => result.push(value),
-                _ => missing_labels.push(label_name),
-            }
-        }
-
-        if missing_labels.is_empty() {
-            Ok(result.join(","))
-        } else {
-            Err(Error::PodWithoutLabelsNotSupported(missing_labels))
-        }
-    }
-
-    pub fn try_from_pod_and_id(pod: &Pod, id: &str) -> SchedulerResult<Self> {
-        if id.is_empty() {
-            return Err(Error::PodIdentityFieldEmpty);
-        }
-
-        match &pod.metadata.labels {
-            Some(labels) => {
-                let mut missing_labels = Vec::with_capacity(4);
-                let mut app = String::new();
-                let mut instance = String::new();
-                let mut role = String::new();
-                let mut group = String::new();
-
-                match labels.get(labels::APP_NAME_LABEL).cloned() {
-                    Some(value) => app = value,
-                    _ => missing_labels.push(labels::APP_NAME_LABEL),
-                }
-                match labels.get(labels::APP_INSTANCE_LABEL).cloned() {
-                    Some(value) => instance = value,
-                    _ => missing_labels.push(labels::APP_INSTANCE_LABEL),
-                }
-                match labels.get(labels::APP_COMPONENT_LABEL).cloned() {
-                    Some(value) => role = value,
-                    _ => missing_labels.push(labels::APP_COMPONENT_LABEL),
-                }
-                match labels.get(labels::APP_ROLE_GROUP_LABEL).cloned() {
-                    Some(value) => group = value,
-                    _ => missing_labels.push(labels::APP_ROLE_GROUP_LABEL),
-                }
-
-                if missing_labels.is_empty() {
-                    Ok(PodIdentity::new(
-                        app.as_str(),
-                        instance.as_str(),
-                        role.as_str(),
-                        group.as_str(),
-                        id,
-                    ))
-                } else {
-                    Err(Error::PodWithoutLabelsNotSupported(missing_labels))
-                }
-            }
-            _ => Err(Error::PodWithoutLabelsNotSupported(
-                REQUIRED_LABELS.to_vec(),
-            )),
-        }
-    }
-
-    pub fn app(&self) -> &str {
-        self.app.as_ref()
-    }
-    pub fn instance(&self) -> &str {
-        self.instance.as_ref()
-    }
-    pub fn role(&self) -> &str {
-        self.role.as_ref()
-    }
-    pub fn group(&self) -> &str {
-        self.group.as_ref()
-    }
-    pub fn id(&self) -> &str {
-        self.id.as_ref()
-    }
-
-    pub fn compute_hash(&self, hasher: &mut DefaultHasher) -> u64 {
-        self.hash(hasher);
-        hasher.finish()
-    }
-
-    fn warn_forbidden_char(app: &str, instance: &str, role: &str, group: &str, id: &str) {
-        if app.contains(SEMICOLON) {
-            warn!(
-                "Found forbidden character [{}] in application name: {}",
-                SEMICOLON, app
-            );
-        }
-        if instance.contains(SEMICOLON) {
-            warn!(
-                "Found forbidden character [{}] in instance name: {}",
-                SEMICOLON, instance
-            );
-        }
-        if role.contains(SEMICOLON) {
-            warn!(
-                "Found forbidden character [{}] in role name: {}",
-                SEMICOLON, role
-            );
-        }
-        if group.contains(SEMICOLON) {
-            warn!(
-                "Found forbidden character [{}] in group name: {}",
-                SEMICOLON, group
-            );
-        }
-        if id.contains(SEMICOLON) {
-            warn!(
-                "Found forbidden character [{}] in pod id: {}",
-                SEMICOLON, id
-            );
-        }
-    }
-}
 impl<'a> HashingStrategy<'a> {
     pub fn new(eligible_nodes: &'a RoleGroupEligibleNodes) -> Self {
         Self {
@@ -945,7 +489,7 @@ impl<'a> HashingStrategy<'a> {
     fn select_node_for_pod(
         &self,
         pod: &PodIdentity,
-        preferred_node: Option<&NodeIdentity>,
+        preferred_node: Option<NodeIdentity>,
     ) -> Option<NodeIdentity> {
         self.eligible_nodes
             .preferred_node_or(pod, preferred_node, |pod, nodes| {
@@ -959,23 +503,25 @@ impl<'a> HashingStrategy<'a> {
 impl PodPlacementStrategy for HashingStrategy<'_> {
     fn place(
         &self,
-        pods: &[&PodIdentity],
-        preferred_nodes: &[Option<&NodeIdentity>],
+        pods: &[PodIdentity],
+        preferred_nodes: &[Option<NodeIdentity>],
     ) -> Vec<Option<NodeIdentity>> {
         assert_eq!(pods.len(), preferred_nodes.len());
         pods.iter()
             .zip(preferred_nodes.iter())
-            .map(|(pod, preferred_node)| self.select_node_for_pod(*pod, *preferred_node))
+            .map(|(pod, preferred_node)| self.select_node_for_pod(pod, preferred_node.clone()))
             .collect()
     }
 }
-
+/*
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::builder::{ObjectMetaBuilder, PodBuilder};
-    use crate::role_utils::EligibleNodesAndReplicas;
     use rstest::*;
+
+    use crate::identity;
+    use crate::role_utils::EligibleNodesAndReplicas;
+
+    use super::*;
 
     const APP_NAME: &str = "app";
     const INSTANCE: &str = "simple";
@@ -991,8 +537,8 @@ mod tests {
     ///     - group_0:
     ///         - id: 1
     ///     - group_1:
-    ///         - id: 1
     ///         - id: 2
+    ///         - id: 3
     ///         - ... # up to `how_many` - 1
     /// ```
     fn generate_ids(how_many: usize) -> Vec<PodIdentity> {
@@ -1023,7 +569,7 @@ mod tests {
             .cloned()
             .collect(),
         );
-        super::generate_ids(APP_NAME, INSTANCE, &eligible_nodes)
+        identity::generate_ids(APP_NAME, INSTANCE, &eligible_nodes, 1)
     }
 
     impl PodPlacementHistory for TestHistory {
@@ -1082,7 +628,7 @@ mod tests {
 
         for pod_id in scheduled_pods {
             let nodes = available_nodes
-                .get_nodes_mut(&pod_id.role, &pod_id.group)
+                .get_nodes_mut(pod_id.role(), pod_id.group())
                 .unwrap();
             current_mapping.insert(pod_id.clone(), nodes.pop().unwrap().clone());
         }
@@ -1224,12 +770,12 @@ mod tests {
     #[case::place_one_pod_on_preferred(2, 0, 5, &[None, Some(NodeIdentity { name: "NODE_2".to_string() })], &[Some(NodeIdentity { name: "NODE_0".to_string() }), Some(NodeIdentity { name: "NODE_2".to_string() })])]
     #[case::place_three_pods(3, 0, 5, &[None, None, None],
         &[Some(NodeIdentity { name: "NODE_0".to_string() }),
-          Some(NodeIdentity { name: "NODE_4".to_string() }),
-          Some(NodeIdentity { name: "NODE_4".to_string() })])]
+          Some(NodeIdentity { name: "NODE_2".to_string() }),
+          Some(NodeIdentity { name: "NODE_1".to_string() })])]
     #[case::place_three_pods_one_on_preferred(3, 0, 5, &[None, Some(NodeIdentity { name: "NODE_3".to_string() }), None],
         &[Some(NodeIdentity { name: "NODE_0".to_string() }),
           Some(NodeIdentity { name: "NODE_3".to_string() }),
-          Some(NodeIdentity { name: "NODE_2".to_string() })])]
+          Some(NodeIdentity { name: "NODE_4".to_string() })])]
     fn test_scheduler_hashing_strategy(
         #[case] wanted_pod_count: usize,
         #[case] scheduled_pods_count: usize,
@@ -1270,11 +816,7 @@ mod tests {
         #[case] expected: Option<NodeIdentity>,
     ) {
         let eligible_nodes = generate_eligible_nodes(eligible_node_count);
-        let pod = PodIdentity {
-            role: role.to_string(),
-            group: group.to_string(),
-            ..PodIdentity::default()
-        };
+        let pod = PodIdentity::new("APP", "INSTANCE", role, group, "0");
         let got = eligible_nodes.preferred_node_or_last(&pod, opt_node_id.as_ref());
 
         assert_eq!(got, expected);
@@ -1293,7 +835,7 @@ mod tests {
 
     #[rstest]
     #[case::no_missing_pods(1, 1, 1, vec![])]
-    #[case::missing_one_pod(1, 0, 1, vec![PodIdentity { app: "app".to_string(), instance: "simple".to_string(), role: "ROLE_0".to_string(), group: "GROUP_0".to_string(), id: "1".to_string() }])]
+    #[case::missing_one_pod(1, 0, 1, vec![PodIdentity::new("app", "simple", "ROLE_0", "GROUP_0", "1")])]
     fn test_scheduler_pod_to_node_mapping_missing(
         #[case] wanted_pod_count: usize,
         #[case] scheduled_pods_count: usize,
@@ -1321,20 +863,14 @@ mod tests {
            current_mapping: PodToNodeMapping::default(),
            remaining_mapping:
                PodToNodeMapping::new(vec![
-                       (PodIdentity {app: "app".to_string(), instance: "simple".to_string(), role: "ROLE_0".to_string(), group: "GROUP_0".to_string(), id: "1".to_string() }, NodeIdentity { name: "NODE_0".to_string() }),
+                       (PodIdentity::new("app", "simple", "ROLE_0", "GROUP_0", "1"), NodeIdentity { name: "NODE_0".to_string() }),
                    ])},
        ))]
     #[case::pod_cannot_be_scheduled(1, 0,
         Err(Error::NotEnoughNodesAvailable {
             number_of_nodes: 0,
             number_of_pods: 1,
-            unscheduled_pods: vec![
-                PodIdentity {
-                    app: "app".to_string(),
-                    instance: "simple".to_string(),
-                    role: "ROLE_0".to_string(),
-                    group: "GROUP_0".to_string(),
-                    id: "1".to_string() }] }))]
+            unscheduled_pods: vec!["PodIdentity { app: \"app\", instance: \"simple\", role: \"ROLE_0\", group: \"GROUP_0\", id: \"1\" }".to_string()] }))]
     fn test_scheduler_update_history_and_result(
         #[case] pod_count: usize,
         #[case] node_count: usize,
@@ -1365,51 +901,12 @@ mod tests {
             &current_mapping,
         );
 
-        assert_eq!(got, expected);
-    }
-
-    #[rstest]
-    #[case(&[], "", Err(Error::PodIdentityFieldEmpty))]
-    #[case(&[], "1", Err(Error::PodWithoutLabelsNotSupported(REQUIRED_LABELS.to_vec())))]
-    #[case::no_app_label(&[(labels::APP_INSTANCE_LABEL, "myinstance"),
-            (labels::APP_COMPONENT_LABEL, "myrole"),
-            (labels::APP_ROLE_GROUP_LABEL, "mygroup")],
-        "2",
-        Err(Error::PodWithoutLabelsNotSupported([labels::APP_NAME_LABEL].to_vec())))]
-    #[case(&[(labels::APP_NAME_LABEL, "myapp"),
-            (labels::APP_INSTANCE_LABEL, "myinstance"),
-            (labels::APP_COMPONENT_LABEL, "myrole"),
-            (labels::APP_ROLE_GROUP_LABEL, "mygroup")],
-        "2",
-        Ok(PodIdentity{
-            app: "myapp".to_string(),
-            instance: "myinstance".to_string(),
-            role: "myrole".to_string(),
-            group: "mygroup".to_string(),
-            id: "2".to_string()}))]
-    fn test_scheduler_pod_identity_try_from_pod_and_id(
-        #[case] labels: &[(&str, &str)],
-        #[case] id: &str,
-        #[case] expected: SchedulerResult<PodIdentity>,
-    ) {
-        let labels_map: BTreeMap<String, String> = labels
-            .iter()
-            .map(|t| (t.0.to_string(), t.1.to_string()))
-            .collect();
-        let pod = PodBuilder::new()
-            .metadata(
-                ObjectMetaBuilder::new()
-                    .generate_name("pod1")
-                    .namespace("default")
-                    .with_labels(labels_map)
-                    .build()
-                    .unwrap(),
-            )
-            .build()
-            .unwrap();
-
-        let got = PodIdentity::try_from_pod_and_id(&pod, id);
-
-        assert_eq!(got, expected);
+        // Cannot compare `SchedulerResult`s directly because `crate::error::Error` doesn't implement `PartialEq`
+        match (&got, &expected) {
+            (Ok(g), Ok(e)) => assert_eq!(g, e),
+            (Err(ge), Err(re)) => assert_eq!(format!("{:?}", ge), format!("{:?}", re)),
+            _ => panic!("got: {:?}\nexpected: {:?}", got, expected),
+        }
     }
 }
+*/
